+++ spec/www/spec/db-tx-sql-features-test.js
@@ -166,10 +166,12 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        // Test for Cordova-sqlcipher-adapter version (SQLCipher 3.4.0 based on SQLite 3.11.0)
+        // XXX TBD (Android plugin ONLY ...)
         it(suiteName + 'Basic JSON1 json test', function(done) {
-          //if (isWebSql) pending('SKIP for Web SQL (not implemented)');
-          pending('SKIP: NOT IMPLEMENTED for this version');
+          if (isWebSql) pending('SKIP for Web SQL (not implemented)');
+          if (isWindows) pending('SKIP: NOT IMPLEMENTED for Windows'); // FUTURE TBD
+          if (!isWebSql && !isWindows && isAndroid && isImpl2) pending('TBD SKIP for androidDatabaseImplementation: 2'); // XXX TBD
+          if (!isWebSql && !isAndroid && !isWindows) pending('SKIP: NOT IMPLEMENTED for iOS/macOS plugin'); // FUTURE TBD
 
           var db = openDatabase('basic-json1-json-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
@@ -196,10 +198,12 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        // Test for Cordova-sqlcipher-adapter version (SQLCipher 3.4.0 based on SQLite 3.11.0)
+        // XXX TBD (Android plugin ONLY ...)
         it(suiteName + 'JSON1 json_object test', function(done) {
-          //if (isWebSql) pending('SKIP for Web SQL (not implemented)');
-          pending('SKIP: NOT IMPLEMENTED for this version');
+          if (isWebSql) pending('SKIP for Web SQL (not implemented)');
+          if (isWindows) pending('SKIP: NOT IMPLEMENTED for Windows'); // FUTURE TBD
+          if (!isWebSql && !isWindows && isAndroid && isImpl2) pending('TBD SKIP for androidDatabaseImplementation: 2'); // XXX TBD
+          if (!isWebSql && !isAndroid && !isWindows) pending('SKIP: NOT IMPLEMENTED for iOS/macOS plugin'); // FUTURE TBD
 
           var db = openDatabase('json1-json-object-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
@@ -227,10 +231,12 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        // Test for Cordova-sqlcipher-adapter version (SQLCipher 3.4.0 based on SQLite 3.11.0)
+        // XXX TBD (Android plugin ONLY ...)
         it(suiteName + 'create virtual table using FTS5', function(done) {
-          //if (isWebSql) pending('SKIP for Web SQL (not implemented)');
-          pending('SKIP: NOT IMPLEMENTED for this version');
+          if (isWebSql) pending('SKIP for Web SQL (not implemented)');
+          if (isWindows) pending('SKIP: NOT IMPLEMENTED for Windows'); // FUTURE TBD
+          if (!isWebSql && !isWindows && isAndroid && isImpl2) pending('TBD SKIP for androidDatabaseImplementation: 2'); // XXX TBD
+          if (!isWebSql && !isAndroid && !isWindows) pending('SKIP: NOT IMPLEMENTED for iOS/macOS plugin'); // FUTURE TBD
 
           var db = openDatabase('virtual-table-using-fts5.db', '1.0', 'Test', DEFAULT_SIZE);