+++ src/ios/SQLitePlugin.m
@@ -10,6 +10,10 @@
 
 #import "sqlite3.h"
 
+#import "sqlite3_regexp.h"
+
+#import "sqlite3_base64.h"
+
 #import "PSPDFThreadSafeMutableDictionary.h"
 
 // Defines Macro to only log lines when in DEBUG mode
@@ -143,6 +147,12 @@
                 [self.commandDelegate sendPluginResult:pluginResult callbackId: command.callbackId];
                 return;
             } else {
+                // TBD IGNORE result:
+                const char * err1;
+                sqlite3_regexp_init(db, &err1);
+
+                sqlite3_base64_init(db);
+
                 // for SQLCipher version:
                 // NSString *dbkey = [options objectForKey:@"key"];
                 // const char *key = NULL;