+++ CHANGES.md
@@ -1,8 +1,8 @@
 # Changes
 
-### cordova-sqlite-evcore-legacy-ext-common-free 0.0.2-dev01
+### cordova-sqlite-evcore-legacy-ext-common-free 0.0.2-dev02
 
-TBD
+- Remove SQLITE_ENABLE_COLUMN_METADATA not needed (iOS/macOS/Windows)
 
 ### cordova-sqlite-evcore-legacy-ext-common-free 0.0.1