+++ CHANGES.md
@@ -1,8 +1,8 @@
 # Changes
 
-### cordova-sqlite-evcore-legacy-ext-common-free 0.0.3-dev01
+### cordova-sqlite-evcore-legacy-ext-common-free 0.0.3
 
-TBD
+- cordova-sqlite-evcore-legacy-ext-common-free with SQLITE_THREADSAFE=1 on iOS/macOS ref: litehelpers/Cordova-sqlite-storage#754 (<https://github.com/litehelpers/Cordova-sqlite-storage/issues/754>)
 
 ### cordova-sqlite-evcore-legacy-ext-common-free 0.0.2